+++ include/Nazara/Network/TcpClient.hpp
@@ -45,6 +45,8 @@ namespace Nz
 			inline bool IsLowDelayEnabled() const;
 			inline bool IsKeepAliveEnabled() const;
 
+			SocketState PollForConnected(UInt64 waitDuration = 0);
+
 			bool Receive(void* buffer, std::size_t size, std::size_t* received);
 			bool ReceivePacket(NetPacket* packet);